# Backend/.choreo/Backend/.choreo/openapi.yaml
openapi: 3.0.3
info:
  title: SRI-KO Learning Management System API
  version: 1.0.0
  description: RESTful API for managing educational services including course management, user authentication, subscription services, payment processing, and administrative features
servers:
  - url: https://sri-ko-lms-api.choreo.dev/
    description: Production server
  - url: https://sri-ko-lms-api-dev.choreo.dev/
    description: Development server
  - url: http://localhost:5000/
    description: Local development server
paths:
...
  /api/courses:
    get:
      summary: Get all courses
      tags: [Courses]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  courses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                  page:
                    type: integer
                  pages:
                    type: integer
                  total:
                    type: integer

    post:
      summary: Create a new course
      tags: [Courses]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Course'
        '401':
          description: Unauthorized
        '400':
          description: Validation error
        '403':
          description: Instructor/Admin access required
...
  /api/admin/users:
    get:
      summary: Get all users
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: role
          in: query
          schema:
            type: string
            enum: [student, instructor, admin]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  page:
                    type: integer
                  pages:
                    type: integer
                  total:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the user
        name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        role:
          type: string
          enum: [student, instructor, admin]
          default: student
        phone:
          type: string
        avatar:
          type: string
          default: default-avatar.jpg
        isActive:
          type: boolean
          default: true
        emailVerified:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [name, email, role]
...
    Course:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the course
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
          default: beginner
        duration:
          type: integer
          minimum: 1
        price:
          type: number
          format: float
          minimum: 0
        instructor:
          type: string
          description: Reference to User ID
        enrolledStudents:
          type: array
          items:
            type: string
        curriculum:
          type: array
          items:
            type: object
            properties:
              week:
                type: integer
              title:
                type: string
              description:
                type: string
              lessons:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                    duration:
                      type: integer
                    isFreePreview:
                      type: boolean
        prerequisites:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        averageRating:
          type: number
          format: float
        reviews:
          type: array
          items:
            type: object
            properties:
              user:
                type: string
              rating:
                type: integer
                minimum: 1
                maximum: 5
              comment:
                type: string
              createdAt:
                type: string
                format: date-time
        isPublished:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [title, description, category, level, duration, price, instructor]
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          minLength: 1
          maxLength: 2000
        category:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        duration:
          type: integer
          minimum: 1
        price:
          type: number
          format: float
          minimum: 0
        curriculum:
          type: array
          items:
            type: object
            properties:
              week:
                type: integer
              title:
                type: string
              description:
                type: string
              lessons:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                    duration:
                      type: integer
                    isFreePreview:
                      type: boolean
        prerequisites:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
      required: [title, description, category, level, duration, price]
...